Rangepick: Use gtk.Popover for the rangepicker popup
Unlike the previously used gtk.Window, gtk.Popover automatically handles
positioning and automatically closes when the user clicks outside of the
popover, so seems more suitable here.

One downside of using a Popover is that it is a widget, rather than an
window, so at least on X11, it cannot extend outside of the containing
window, but this is probably not a problem in practice (this limitation
does not seem to exist on Wayland, see
https://gitlab.gnome.org/GNOME/gtk/-/issues/543).

The popover also autocloses when clicking outside of the popover. For
this, it grabs the input, so the click that closes the popover does not
trigger any other actions. This allows removing some code on focus out
that prevent the popup from re-opening when clicking on the button to
close the popover, which conveniently also fixes #639.

Switching to a popover also seems to fix positioning of the popup on
wayland (which may have been caused by a bug in gtk, or by improper
setup of the popup window on the Hamster side, but are conveniently
sidestepped by this change in any case). This fixes #645.